# docs/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: This is a rule engine server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rule Engine API
  version: "1.0"
paths:
  /api/v1/actions:
    get:
      description: Get a list of all actions with optional pagination.
      parameters:
        - description: Limit number of actions returned
          in: query
          name: limit
          schema:
            type: integer
        - description: Offset for pagination
          in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: List all actions
      tags:
        - actions
    post:
      description: Create a new action with the given name and Lua script.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.CreateActionRequest"
        description: Action to create
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ActionInfo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Create a new action
      tags:
        - actions
  "/api/v1/actions/{id}":
    get:
      description: Get a single action by its unique ID.
      parameters:
        - description: Action ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ActionInfo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Get an action by ID
      tags:
        - actions
    patch:
      description: Update an existing action using a JSON Patch.
      parameters:
        - description: Action ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/api.PatchOperationArray"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ActionInfo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Update an action
      tags:
        - actions
...
  "/api/v1/rules/{id}/actions":
    post:
      description: Add an existing action to an existing rule.
      parameters:
        - description: Rule ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.AddActionToRuleRequest"
        description: Action to add
        required: true
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Add an action to a rule
      tags:
        - rules
...
servers:
  - url: //localhost:8080/api/v1
components:
  requestBodies:
    api.PatchOperationArray:
      content:
        application/json:
          schema:
            items:
              $ref: "#/components/schemas/api.PatchOperation"
            type: array
      description: JSON Patch operations
      required: true
  schemas:
    api.APIErrorResponse:
      properties:
        error:
          example: Error message
          type: string
      type: object
    api.ActionInfo:
      properties:
        created_at:
          type: string
        enabled:
          type: boolean
        id:
          type: string
        lua_script:
          type: string
        name:
          type: string
        updated_at:
          type: string
      type: object
    api.AddActionToRuleRequest:
      properties:
        action_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
      required:
        - action_id
      type: object
    api.CreateActionRequest:
      properties:
        enabled:
          example: true
          type: boolean
        lua_script:
          example: log_message('info', 'Temperature alert triggered')
          type: string
        name:
          example: Send Temperature Alert
          type: string
      required:
        - lua_script
      type: object
...
    api.PatchOperation:
      properties:
        op:
          enum:
            - add
            - remove
            - replace
            - test
          example: replace
          type: string
        path:
          example: /name
          type: string
        value: {}
      required:
        - op
        - path
      type: object
...